Update README.md adding some of Wayne's onboarding experience
I have added a bit more detail to the readme based on my experience getting setup. Since not everyone uses 1Password, I have not added any instructions relating to that.  As they stand, these instructions may be still a little esoteric for a newbie, but given the mostly singleton nature of this repo, I assume anybody using the readme is most likely to be in contact with the EOP team anyway.